Added Anti-Cheat
Currently has no repercussions, but will show error if shop.txt is tampered with.

Uses SHA-512 to hash the sum of the numbers, plus a secret key string. Compares the pre-generated, stored hash with one derived from the numbers, and checks if they agree.
Obviously, since the source code is vieweable, anyone with knowledge of hashing or python will be able to reverse-egineer it, but anyone else will not be able to derive the secret key from the hash, nor change anything without looking at the source code.